Fixes
Different minor fixes

* Removed unneeded Bitvavo table
* Fixed warning while loading the app
